Refactored files (partially into new ones): - `carapace.go` has stripped from completion and command functions. - `context` has a new `newContext` function - `command.go` contains the function to bind the carapace command
More detailed explanations are given in the PR comment.
